+++ backend/src/main/java/team6/codesquad/kiosk/order/controller/OrderController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
-import team6.codesquad.kiosk.order.controller.dto.OrderRequest;
+import team6.codesquad.kiosk.order.dto.request.OrderRequestDto;
 import team6.codesquad.kiosk.order.service.OrderService;
 
 @RequiredArgsConstructor
@@ -16,8 +16,8 @@ public class OrderController {
 	private final OrderService service;
 
 	@PostMapping("/order")
-	public ResponseEntity<Void> createOrder(@RequestBody OrderRequest orderRequest) {
-		service.saveOrder(orderRequest);
+	public ResponseEntity<Void> createOrder(@RequestBody OrderRequestDto orderRequestDto) {
+		service.saveOrder(orderRequestDto);
 		return ResponseEntity.ok().build();
 	}
 }
