+++ backend/src/main/java/team6/codesquad/kiosk/order/dto/request/MenuRequestDto.java
@@ -1,4 +1,4 @@
-package team6.codesquad.kiosk.order.controller.dto;
+package team6.codesquad.kiosk.order.dto.request;
 
 import java.util.List;
 
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class MenuDto {
+public class MenuRequestDto {
 	private String name;
 	private int count;
 	private List<String> options;
