+++ backend/src/main/java/team6/codesquad/kiosk/order/service/OrderService.java
@@ -6,12 +6,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
-import team6.codesquad.kiosk.order.controller.dto.MenuDto;
-import team6.codesquad.kiosk.order.controller.dto.OrderRequest;
 import team6.codesquad.kiosk.order.domain.Order;
 import team6.codesquad.kiosk.order.domain.OrderMenu;
 import team6.codesquad.kiosk.order.domain.OrderMenuOption;
 import team6.codesquad.kiosk.order.domain.Sales;
+import team6.codesquad.kiosk.order.dto.request.MenuRequestDto;
+import team6.codesquad.kiosk.order.dto.request.OrderRequestDto;
 import team6.codesquad.kiosk.order.repository.MenuRepository;
 import team6.codesquad.kiosk.order.repository.OptionRepository;
 import team6.codesquad.kiosk.order.repository.OrderMenuOptionRepository;
@@ -31,28 +31,28 @@ public class OrderService {
 	private final OrderMenuOptionRepository menuOptionRepository;
 	private final OptionRepository optionRepository;
 
-	public void saveOrder(OrderRequest orderRequest) {
-		Order order = getOrder(orderRequest);
+	public void saveOrder(OrderRequestDto orderRequestDto) {
+		Order order = getOrder(orderRequestDto);
 
 		int orderId = orderRepository.save(order);
-		List<MenuDto> menus = orderRequest.getMenus();
-		for (MenuDto menu : menus) {
+		List<MenuRequestDto> menus = orderRequestDto.getMenus();
+		for (MenuRequestDto menu : menus) {
 			int orderMenuId = saveOrderMenu(orderId, menu);
 			saveOrderMenuOption(orderMenuId, menu);
 		}
 	}
 
-	private Order getOrder(OrderRequest orderRequest) {
-		int totalAmount = orderRequest.getTotalAmount();
-		int receivedAmount = orderRequest.getReceivedAmount();
+	private Order getOrder(OrderRequestDto orderRequestDto) {
+		int totalAmount = orderRequestDto.getTotalAmount();
+		int receivedAmount = orderRequestDto.getReceivedAmount();
 		int change = receivedAmount - totalAmount;
 
 		return new Order(totalAmount, receivedAmount, change,
-			paymentRepository.findByName(orderRequest.getPaymentType()));
+			paymentRepository.findByName(orderRequestDto.getPaymentType()));
 
 	}
 
-	private int saveOrderMenu(int orderId, MenuDto menu) {
+	private int saveOrderMenu(int orderId, MenuRequestDto menu) {
 		int menuId = menuRepository.findByName(menu.getName());
 		int count = menu.getCount();
 		OrderMenu orderMenu = new OrderMenu(orderId, menuId, count);
@@ -62,7 +62,7 @@ public class OrderService {
 		return orderMenuId;
 	}
 
-	private void saveOrderMenuOption(int orderMenuId, MenuDto menu) {
+	private void saveOrderMenuOption(int orderMenuId, MenuRequestDto menu) {
 		for (String option : menu.getOptions()) {
 			OrderMenuOption menuOption = new OrderMenuOption(orderMenuId, optionRepository.findByName(option));
 			menuOptionRepository.save(menuOption);
